freitas desliga os swap files no vim, são lixo
usa undo files

+++ src/ea.c
@@ -151,14 +151,15 @@ void count_strategy(Player *player, Card *new_card, Megadeck *megadeck)
 		megadeck->count++;
 	else if (new_card->id > 7)
 		megadeck->count--;
-	
+
 	player->count = megadeck->count;
 }
 
+// TODO: lá vamos nós falar com o ascenso outra vez
 void hi_lo(List *players, Megadeck *megadeck)
 {
-	int decks_left = (int) (megadeck->cards_left)/DECK_SIZE;
-	int modifier = (int) (megadeck->count)/decks_left;
+	int decks_left = megadeck->cards_left/DECK_SIZE;
+	int modifier = megadeck->count/decks_left;
 
 	List * aux = players->next;
 	while (aux) {
@@ -166,7 +167,7 @@ void hi_lo(List *players, Megadeck *megadeck)
 		if (cur_player->type == EA && modifier > 0)
 			cur_player->bet = cur_player->orig_bet * (modifier+1);
 		aux = aux->next;
-	}		
+	}
 }
-		 
+
